requirements.txt
# AI Railway Wagon Detection System - Requirements
# Python 3.8+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0

# Object Detection
ultralytics>=8.0.0

# OCR
easyocr>=1.7.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Image Quality Metrics
pytorch-msssim>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Documentation
python-docx>=0.8.11
